Fixed test and moved test to better place
