+++ Camunda.Api.Client.Tests/ProcessInstanceTests.cs
@@ -16,29 +16,14 @@ namespace Camunda.Api.Client.Tests
         {
             var mockHttp = new MockHttpMessageHandler();
 
-            mockHttp.Expect(HttpMethod.Post, "http://localhost:8080/engine-rest")
-                .Respond(HttpStatusCode.OK, "text/html", "OK");
+            mockHttp.Expect(HttpMethod.Post, "http://localhost:8080/engine-rest/process-instance")
+                .Respond(HttpStatusCode.OK, "application/json", "[]");
 
 
             var client = CamundaClient.Create("http://localhost:8080/engine-rest", mockHttp);
             var process = await client.ProcessInstances.Query().List();
 
             Assert.NotNull(process);
-        }
-
-        [Fact]
-        public async Task GetVersion()
-        {
-            var mockHttp = new MockHttpMessageHandler();
-
-            mockHttp.Expect(HttpMethod.Get, "http://localhost:8080/engine-rest/version")
-                .Respond(HttpStatusCode.OK, "text/html", "{\"version\": \"7.14.0\"}");
-
-
-            var client = CamundaClient.Create("http://localhost:8080/engine-rest", mockHttp);
-            var version = await client.Version.Get();
-
-            Assert.Equal("7.14.0", version.Version);
         }
     }
 }
